Puntuación y arreglos
* Añadida la puntuación al juego.
* Añadido dos textos a la interfaz: una etiqueta del jugador y uno para
la puntuación.
* Añadidos los pickups al juego.
* Ahora los enemigos sueltan pickups al morir.
* Añadido el ojo del dinosaurio animado.
* Ahora el collider exterior del jugador también se considera de tipo
PLAYER.
* Añadido un método al PlayerLifeComponent para restaurar vida.
* Añadidos más enemigos a la escena.
* Añadido el color predefinido negro.
* Ahora las entidades hacen update y demás a sus componentes antes que a
sus hijos.
* Ahora la posición por defecto de los elementos de GUI es TOP_LEFT en
vez de ABSOLUTE.
* Ahora el GUITextRendererComponent tiene un alineamiento que permite
dibujar el texto hacia un lado u otro del punto de referencia.
* Solucionado un problema que ocasionaba que los colliders no se
dibujasen correctamente en modo DEBUG.

+++ CavemanNinja/GUILifebarComponent.cpp
@@ -33,7 +33,7 @@ bool GUILifebarComponent::OnStart()
 	}
 
 	// Establece los puntos de vida representados
-	representedLifePoints = lifeComponent->currentLifePoints;
+	representedLifePoints = (float) lifeComponent->currentLifePoints;
 
 	return true;
 }
